secure type controller

+++ dashboard/app/Http/Controllers/TypeController.php
@@ -4,15 +4,17 @@ use Illuminate\Http\Response;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use App\Util\MtHttp;
+use App\Util\Access;
 
 class TypeController extends Controller
 {
-	public function __construct(Request $request) {
+    public function __construct(Request $request) {
 		$this->middleware('auth');
-  }
+    }
 
 	public function index(Request $request, $appid)
 	{
+        if(Access::can_view($request->user()->id, $appid) == false) abort(500,'Permission Denied');
 		$result = MtHttp::get('actiontype/' . $appid);
 		return view('actiontype/index', [
 			'appcode' => $appid,
@@ -22,6 +24,7 @@ class TypeController extends Controller
 
 	public function show(Request $request, $appid, $id)
 	{
+        if(Access::can_view($request->user()->id, $appid) == false) abort(500,'Permission Denied');
 		$result = MtHttp::get('actiontype/' . $appid . '/' . $id);
 		if($result){
 			return view('actiontype/edit', [
@@ -35,6 +38,7 @@ class TypeController extends Controller
 	
 	public function update(Request $request, $appid, $id)
 	{
+        if(Access::can_editStruct($request->user()->id, $appid) == false) abort(500,'Permission Denied');
 		$p = $request->all();
 		$p['fields'] = array();
 		for($i = 0; $i < count($p['pcodes']); $i++){
@@ -55,10 +59,12 @@ class TypeController extends Controller
 
 	public function getcreate(Request $request, $appcode)
 	{	
+ if(Access::can_editStruct($request->user()->id, $appcode) == false) abort(500,'Permission Denied');
 		return view('actiontype/create', ['appcode' => $appcode]);
 	}
 
 	public function store(Request $request, $appcode){
+ if(Access::can_editStruct($request->user()->id, $appcode) == false) abort(500,'Permission Denied');
 		$p = $request->all();
 		$p['fields'] = array();
 		for($i = 0; $i < count($p['pcodes']); $i++){
@@ -80,6 +86,7 @@ class TypeController extends Controller
 
 	public function destroy(Request $request, $appcode, $id)
 	{
+ if(Access::can_editStruct($request->user()->id, $appcode) == false) abort(500,'Permission Denied');
 		$result = MtHttp::delete('actiontype/' . $appcode . '/' . $id, array());
 		return redirect("actiontype/$appcode");
 	}
